refactor: extract useEntryDetail hook and shared SectionHeader
Both detail components duplicated an identical ~80-line save-timer
state machine (activeField, tempValue, saveStatus, deleteConfirming,
deleteError, two timer refs, and all handlers). SectionHeader was
defined identically across four components. Extracting both eliminates
the need to apply any save/delete UX change in multiple places.

Closes #13
